+++ ChangeLog.md
@@ -2,9 +2,11 @@
 
   * **Note**: breaking change.
     * `Zydis.Decoder.decodeFullBuffer` now return a `Seq` instead of a `Vector`.
+    * introduce `ZyanStatus`, `ZyanCoreStatus` and `ZydisStatus` in `Zydis.Status` which are directly mapped from their C counterparts.
+
   * remove `vector` dependency
   * introduce `containers` dependency
 
 [0.1.1.0] - November 2020
 
-  * Initial release
+  * initial release